% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparenciaBR.R
\name{coletaConvenioGeral}
\alias{coletaConvenioGeral}
\title{Transferência de recursos para os municípios}
\usage{
coletaConvenioGeral()
}
\description{
Essa função extrai os valores dos convênios transferidos aos estados
}
