% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparenciaBR.R
\name{TransferenciaEstado}
\alias{TransferenciaEstado}
\title{Transferência de recursos para os estados}
\usage{
TransferenciaEstado(ano2)
}
\arguments{
\item{ano}{A partir de 2004.}
}
\description{
Essa função extrai todas as transferências constitucionais para todos os estados
presentes no Portal Transparência
}
