% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparenciaBR.R
\name{TransferenciaMunicipio}
\alias{TransferenciaMunicipio}
\title{Transferência de recursos para os municípios}
\usage{
TransferenciaMunicipio(ano, uf)
}
\arguments{
\item{ano}{A partir de 2004.}

\item{uf}{Sigla dos estados.}
}
\description{
Essa função extrai todas as transferências constitucionais para os municípios
presentes no Portal Transparência
}
