% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparenciaBR.R
\name{ColetaConvenioGeral}
\alias{ColetaConvenioGeral}
\title{Transferência de recursos para os municípios}
\usage{
ColetaConvenioGeral()
}
\description{
Essa função extrai os valores dos convênios transferidos aos estados
}
